requirements.txt
asgiref==3.5.2
distlib==0.3.4
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
filelock==3.7.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.15.1
virtualenvwrapper-win==1.2.7
